requirements.txt
anyjson==0.3.3
Authomatic==0.0.13
argparse==1.2.1
beautifulsoup4==4.4.0
bz2file==0.98
celery==3.1.18
decorator==4.0.2
Django==1.8.2
django-dotenv==1.4.1
django-filter==0.11.0
django-pgjson==0.2.3
djangorestframework==3.2.4
feedparser==5.2.1
funcy==1.5
future==0.14.3
futures==3.0.3
gensim==0.11.1
google-api-python-client==1.4.2
gunicorn==19.3.0
httplib2==0.9.2
jsonschema==2.4.0
isodate==0.4.6
lxml==3.4.4
networkx==1.10
nltk==3.0.4
nose==1.3.7
numexpr==2.4.3
numpy==1.11.1
oauth2client==1.5.1
oauthlib==0.7.2
oatuh==1.0.1
openpyxl==2.3.0
psycopg2==2.6
pyLDAvis==1.2.0
py-trello==0.5.0
PyGithub==1.26.0
python-igraph==0.7.1.post6
python-memchached==1.53
redis==2.10.3
requests==2.7.0
requests-oauthlib==0.5.0
scikit-bio==0.2.3
scikit-learn==0.16.1
scipy==0.15.1
tincan==0.0.5
twython==3.2.0
vaderSentiment==0.5
